requirements.txt
pytest==3.7.3
pytest-allure-adaptor
requests==2.21.0
